requirements.txt
attrs==19.3.0
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
Fiona==1.8.9.post2
geopandas==0.6.1
munch==2.5.0
numpy==1.17.3
pandas==0.25.2
passlib==1.7.1
psycopg2==2.7.5
psycopg2-binary==2.8.3
PyJWT==1.7.1
pyproj==2.4.0
python-dateutil==2.8.0
pytz==2019.2
selenium==3.141.0
Shapely==1.6.4.post2
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
